+++ src/components/Dashboard/Measurements/Add.vue
@@ -8,12 +8,12 @@ const router = useRouter();
 
 const addMeasurement = (measurement) => {
   measurementStore.add(measurement);
-  router.push({name: 'Measurements'});
+  router.back();
 }
 </script>
 
 <template>
-  <Dialog :open="true" @close="$router.push({ name: 'Measurements' })" class="relative z-50">
+  <Dialog :open="true" @close="$router.back()" class="relative z-50">
     <div class="fixed inset-0 bg-black/30 backdrop-blur-sm" />
     <div class="fixed flex w-screen h-screen top-10 items-start md:items-center justify-center">
       <DialogPanel class="bg-white w-full max-w-xs rounded-2xl shadow-lg">