+++ src/components/Dashboard/Measurements/Form.vue
@@ -61,7 +61,7 @@ const vital = computed(() => {
     </label>
     <input v-model="date" type="date" id="date" required>
     <div class="grid grid-flow-col justify-end items-center gap-5 mt-4">
-      <RouterLink :to="{ name: 'Measurements' }" class="text-sm text-gray-500 font-light">Cancel</RouterLink>
+      <a @click="$router.back()" class="text-sm text-gray-500 font-light cursor-pointer">Cancel</a>
       <button type="submit" class="btn" :disabled="!isFormComplete">Save</button>
     </div>
   </form>