+++ CHANGELOG.md
@@ -4,6 +4,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.2] - 2023-09-16
+
+### Added
+- Added placeholder examples for new Vital form
+- Added methods for creating sample vitals for body weight, heart rate
+- Informative notice for no data added to Measurements page
+- Informative notice for no data added to Person page
+- Informative notice for no data added to Vitals page
+- Add measurement route added for Person
+
+### Changed
+- Added documentation notes to stores
+- Add measurement form add action and cancel lead to previous route
+
+### Fixed
+- Vital chart on Person page more reactive with computed data
+
 ## [0.1.1] - 2023-09-15
 
 ### Added